+++ CHANGELOG.md
@@ -34,6 +34,9 @@ Increment the:
   to compile standalone on newer standard library implementations.
   [#4574](https://github.com/open-telemetry/opentelemetry-cpp/pull/4574)
 
+* [SDK] Optimize composite and probability samplers
+  [#4582](https://github.com/open-telemetry/opentelemetry-cpp/pull/4582)
+
 ## [1.29.0] 2026-09-13
 
 * [RELEASE] Bump main branch to 1.29.0-dev (#4259)